move `api/annotation` to `spring-core/annotation`
