+++ elasticjob-lite/elasticjob-lite-spring/elasticjob-lite-spring-core/src/test/java/org/apache/shardingsphere/elasticjob/lite/spring/core/annotation/job/CustomJob.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,28 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.annotation;
+package org.apache.shardingsphere.elasticjob.lite.spring.core.annotation.job;
 
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
+import org.apache.shardingsphere.elasticjob.api.ElasticJob;
+import org.apache.shardingsphere.elasticjob.api.ShardingContext;
 
-/**
- * The annotation that specify elastic-job prop.
- */
-@Documented
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ElasticJobProp {
+public interface CustomJob extends ElasticJob {
     
     /**
-     * prop key.
-     * @return key
+     * Execute custom job.
+     *
+     * @param shardingContext sharding context
      */
-    String key();
-    
-    /**
-     * prop value.
-     * @return value
-     */
-    String value() default "";
+    void execute(ShardingContext shardingContext);
 }
